requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.0.76
mediapipe==0.10.8
scikit-learn==1.3.0
pandas==2.0.3
seaborn==0.12.2
matplotlib==3.7.2
Pillow==10.0.0
streamlit==1.31.0
blinker==1.6.2 